don't bother testing 1.6
I have not been able to figure out the severe flakiness in Go 1.6. And I'm tired of digging for it. Go 1.6 is two years old now, so I suppose it's not valuable to support it. (I had added to the matrix originally just to match the actual grpc-go project, which also supports Go 1.6.)

It's always the same issue: `TestPlainText` in `tls_settings_test.go`. And the behavior appears to be as if the server keeps closing the connection. But it doesn't plague any other test case, which is very mysterious. It's as if the server were configured for TLS and closes the connection when it sees the HTTP/2 preface (which doesn't look like a TLS handshake). But I just don't see how that could be possible, so it must be something else going awry.