allow period at the start of a symbol
When describing a service, InputType and OutputType symbols are presented
as .Foo.Bar. However, when describing a type, .Foo.Bar is not accepted;
it must be changed to Foo.Bar. This is both confusing and breaks copy
paste ability. This commit addresses the problem by optionally allowing a
period at the beginning of a symbol.

Fixes #39.