add goreleaser
Add a goreleaser configuration file that builds binaries for Linux,
MacOS and Windows, for 32bit (x86/386) and 64bit (x64/amd64). The
binaries will be archived into a tar.gz/zip file along with the LICENSE
file.

The `dist/` directory will be written to by goreleaser with the binaries
during the build process, so it's also been added to .gitignore.

Goreleaser can be installed with:
go get github.com/goreleaser/goreleaser

Or other methods for install can be found at:
https://goreleaser.com

To build all the binaries and release them to GitHub:
1. Tag the release. e.g. `git tag -a v1.0.0 -m 'First release'`
2. Generate a GitHub Personal Access Token. See https://github.com/settings/tokens.
3. Push the release to GitHub. e.g. `git push origin v1.0.0`
4. Run goreleaser. e.g.  `GITHUB_TOKEN=xxxxxxx... goreleaser`

Go releaser will build the binaries, and upload the binaries to the
GitHub project's release page for the release that was tagged.